Move `egg-moves.ts`
