+++ src/ui/components/pokemon-hatch-info-container.ts
@@ -4,7 +4,7 @@ import { ElementalType } from "#enums/elemental-type";
 import { rgbHexToRgba, leftPad } from "#app/utils";
 import { addTextObject } from "#app/ui/text/text-utils";
 import { TextStyle } from "#enums/text-style";
-import { speciesEggMoves } from "#app/data/balance/egg-moves";
+import { speciesEggMoves } from "#app/data/egg-moves";
 import { allMoves } from "#app/data/data-lists";
 import { SpeciesId } from "#enums/species-id";
 import { getEggTierForSpecies } from "#app/data/egg";
